#7569 Move the Timer class into a new Piwik\Timer namespace + tag it as API
